requirements.txt
alpaca-py>=3.3.0
redis>=5.0.1
loguru>=0.7.2
aiohttp>=3.11.11
pandas>=2.2.0
numpy>=1.26.0
ta-lib==0.4.28
python-dotenv>=1.0.1
pytz>=2025.1
requests>=2.32.3
gunicorn>=22.0.0